Fixes ftplugin bug, :FnlSwiss etc, docs
In [neo]vim, &runtimepath already contains after/ directories.
Therefore, do not search for .fnl files in after/ directories. Simply
search for .fnl files in all directories in &runtimepath. For any
after/**/*.fnl, I was loading it twice basically.

I also added :FnlSwiss and allowed all commands (:Fnl, :FnlFiles,
:FnlLines, and :FnlSwiss) to redirect the result to a register, a file,
a vim or lua variable, or the current buffer. This uses syntax like from
:redir with a few extras (`@_` for black hol register, `->` and `->>`
for storing and appending to a global lua variable, and `==` to :put the
result). I parsed the redir stuff with lpeg.