adds elin module, caching, fennel docs
This commit adds the elin module, which has the following functions (and
their companion functions for lua):
- `setup`
- `get-version`
- `caching-enabled?`
- `enable-caching`
- `disable-caching`
- `dofile`
- `get-cache-path`
- `get-uncache-path`
- `get-cache-dir`

**elin module:**

Mostly, the elin module provides mechanisms to allow caching. This
module will be expanded in the future...

very soon: the elin module will allow an init.lua to call
`(elin.startup)`, which will do everything `plugin/startup.fnl` does and
set its guard to prevent it from loading. This will allow bootstrapping
everything in init.lua instead of init.fnl, which will let the user
disable builtin plugins. This is because neovim loads init.lua, then
builtin plugins, then sources `plugin/**/*.lua` files, which means we
can't prevent any from loading. This isn't a big deal in my opinion, but
it's nice to have the flexibility for those who *really* want to disable
those plugins before they load. There isn't a performance cost. They
each take a few hundreds of a millisecond to load at most.

**caching:**

When caching is enabled, elin with auto-cache both modules imported
with `require()` and modules loaded with `(elin.dofile)` as well as
auto-detect and load these cached files. They're stored in
`~/.cache/nvim/elin` by default. This can be changed with
`(elin.setup)`, which is currently the only option `setup` allows. This
uses `string.dump()` to store luajit byte code in a binary `.luac` file.
Loading this is roughly 5 times faster.

You can retrieve the cache dir with `(elin.get-cache-dir)`. You can get
a file's cached path with `(elin.get-cache-path)` and reverse this with
`(elin.get-uncache-path)`. These should be absolute paths, but they
don't have to be.

You can see if caching is enabled with `(elin.caching-enabled?)`.
Because it's experimental, you need to put `(elin.enable-caching)` in
your `init.fnl` (or `init.lua`). It can be disabled with
`(elin.disable-caching)`, but that doesn't make sense.

**fennel docs:**

Technomancy, the current maintainer of fennel, graciously permitted the
docs to be reformated to comply with vim docs specifications, and they
are now distributed (most of them) with this plugin. Now, the user can
do `:help fennel-reference.txt` (etc) to see the docs in neovim. Many
additional tags were carefully added. This may need to be pruned later.
It feels overwhelming and somewhat inconsistent.

**etc:**

The help docs include the fennel documentation, and information about
the elin module has been added to elin.txt and the README.