fix fennel.luac problem
This isn't compatible between different luajit versions. My fennel.luac
is not portable. Therefore, this commit fixes that by caching it on the
first load.

The user can't consent to this. It would be pretty annoying to have to
check because you can't load fennel in order to load their init.fnl, but
we want them to consent in init.fnl. It's a cart before the horse thing,
and if it breaks, it breaks. But it is a +5x performance boost (2ms
instead of 10ms), so I'm going to keep it for now.

Sue me.