do-startup, better docs, init.lua support
**do-startup**

Instead of putting all of the startup into plugin/startup.lua,
`(elin.do-startup)` now has all of the startup log and guards itself
against re-invocation. It accepts options, including `force` to force
startup again even if it already happened and `do-init-fnl` to load an
init.fnl even if there is an init.lua or init.vim.

**init.lua support**

This is very useful because now, the user can have an init.lua and use
`local elin = require("elin"); elin.doStartup { doInitFnl = true }` to
force init.fnl to load in addition to the init.lua.

**better docs**

I rewrote the docs for the elin module. It's a lot clearer now I think.