REPL strings w newlines, tweaks
elin.repl was failing when a line in `onValues` had a newline.
`nvim_buf_set_lines` hates newlines and returns an error when it
encounters them.

Also, removes defunct lua/elin/api.lua, which has no fennel counterpart
and was simply a remnant from an earlier compilation.

The readme is a bit more clear. #Installation has a note at the bottom
that explains that you don't need to have an init.fnl or init.lua and
can simply use elin because it bootstraps itself, and #About before
mentions that better docs are available in the vim docs at
docs/elin.txt.

Vim docs also mention keybindings in FnlRepl: `<C-l>` to "clear screen"
(although the scrollback is intentionally preseved) and `<C-d>` on an
empty line to close the REPL.